# Conflicts:
#   src/main/kotlin/pl/umk/mat/locals/models/User.kt
#   src/main/kotlin/pl/umk/mat/locals/services/AdministratorService.kt
#   src/main/kotlin/pl/umk/mat/locals/services/OfferService.kt

+++ src/main/kotlin/pl/umk/mat/locals/services/UserService.kt
@@ -4,6 +4,7 @@ import com.google.api.client.googleapis.auth.oauth2.GoogleAuthorizationCodeToken
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken
 import com.google.api.client.http.javanet.NetHttpTransport
 import com.google.api.client.json.jackson2.JacksonFactory
+import org.springframework.data.repository.findByIdOrNull
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.crypto.password.PasswordEncoder
@@ -14,6 +15,7 @@ import pl.umk.mat.locals.dto.`in`.*
 import pl.umk.mat.locals.dto.out.*
 import pl.umk.mat.locals.exceptions.BadRequest
 import pl.umk.mat.locals.exceptions.ResourceAlreadyExistException
+import pl.umk.mat.locals.exceptions.ResourceNotFoundException
 import pl.umk.mat.locals.exceptions.UserAuthException
 import pl.umk.mat.locals.models.GuideRequest
 import pl.umk.mat.locals.models.TemporaryUser
@@ -154,7 +156,7 @@ class UserService(
                 email = user.email,
                 telephone = user.telephone,
                 experience = user.experience,
-                avatar = config.imageServerHost + user.avatar,
+                avatar = user.avatar,
                 interests = user.interests.map {
                     InterestDto(it)
                 },
@@ -164,6 +166,10 @@ class UserService(
 
     fun findUserByEmail(email: String): User? {
         return userRepository.findUserByEmail(email)
+    }
+
+    fun findUserById(id: Long): User {
+        return userRepository.findByIdOrNull(id) ?: throw ResourceNotFoundException("User not founded.")
     }
 
     private fun getGoogleAccountInfo(code: String, requestUrl: String): GoogleIdToken.Payload {
@@ -190,7 +196,7 @@ class UserService(
     }
 
     fun getSelfUserInfo(user: User): UserSelfInfo {
-        return UserSelfInfo(user, config.imageServerHost)
+        return UserSelfInfo(user)
     }
 
     fun confirmEmail(emailConfirmationCode: EmailConfirmationCode) {
@@ -273,8 +279,8 @@ class UserService(
 
         val filePatch = config.imageDir + filename
         if (File(filePatch).exists()) throw RuntimeException("File exist try later.")
-
-        File(config.imageDir + user.avatar.substringAfterLast("/")).delete()
+        if (user.avatar.substringAfterLast("/") != "avatar_default.jpg")
+            File(config.imageDir + user.avatar.substringAfterLast("/")).delete()
 
         userRepository.save(
                 user.copy(
@@ -301,8 +307,7 @@ class UserService(
                         experience = newUserData.experience,
                         interests = interestRepository.findAllById(newUserData.interests).asSequence().toList(),
                         gender = newUserData.gender
-                )),
-                config.imageServerHost
+                ))
         )
     }
